Renamed DOT.API to DotAPI
