+++ src/main/java/DotAPI/DotFileParser.java
@@ -1,11 +1,10 @@
-package DOT.API;
+package DotAPI;
 
 // Importing required classes
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.HashMap;
-import java.lang.String.*;
 
 
 /**
