+++ src/main/java/DotAPI/GraphEquivalency.java
@@ -1,4 +1,4 @@
-package DOT.API;
+package DotAPI;
 
 public class GraphEquivalency {
