+++ src/main/java/DotAPI/GraphItem.java
@@ -1,4 +1,4 @@
-package DOT.API;
+package DotAPI;
 
 import java.util.LinkedList;
