+++ src/main/java/DotAPI/Main.java
@@ -1,0 +1,15 @@
+package DotAPI;
+
+import SootUp.CHAAnalysis;
+
+import java.util.Optional;
+
+public class Main {
+
+    public static void main(String[] args) {
+
+        DotToHtmlGenerator.embedDotInHtml("basicTests/basicGraph.dot", "basicTests/HtmlTest.html");
+
+    }
+
+}
